Explicitly delete artefacts in `DeleteMachine` function and do not rely on garbage collection (#6)
* Add a function for creating the `*corev1.Pod` object

On the way, this fixed a bug in the `GetMachineStatus` function.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add a function for creating the `*corev1.Secret` object

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Do not rely on garbage collector for `Pod` deletion

Otherwise, the `Machine` object might already disappear from the system while the "backing VM" (aka the `Pod`) still exists. This can lead to interesting situations (e.g. we saw that the `Node` of a deleted machine was re-registered after deletion).

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Do not rely on garbage collector for `Secret` deletion

Not strictly necessary since the secret could safely be deleted by the garbage collector without any side effects, though still implemented like this for consistency/symmetry reasons.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

Co-authored-by: Tim Ebert <tim.ebert@sap.com>